pyproject.toml
[project]
name = "doctor-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "twilio>=9.7.1",
    "uvicorn>=0.35.0",
]
